Aula da tarde
- GUIs
- Maq estados principal (App) e etcs
- Semaforo "haTrabalho" nas maq de estados que garante espera passiva
- angulo é float e não int

+++ src/tp2/CurvarDireita.java
@@ -3,9 +3,9 @@ package tp2;
 public class CurvarDireita extends Mensagem {
 
 	int raio;
-	int angulo;
+	float angulo;
 	
-	public CurvarDireita(int raio, int angulo) {
+	public CurvarDireita(int raio, float angulo) {
 		super(TIPO_MENSAGEM.CURVA_DIR, String.format("Curva à Direita (raio %d, ângulo %d) \n", raio, angulo));
 		this.raio = raio;
 		this.angulo = angulo;
@@ -14,7 +14,8 @@ public class CurvarDireita extends Mensagem {
 	public int getRaio() {
 		return this.raio;
 	}
-	public int getAngulo() {
+
+	public float getAngulo() {
 		return this.angulo;
 	}
 }
